lexicaforge/api/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
import strawberry
from strawberry.fastapi import GraphQLRouter
from typing import List, Optional
import os

from lexicaforge.db.config import get_db
from lexicaforge.api.schemas.etymology import Query, Mutation

# Create Strawberry Schema
schema = strawberry.Schema(query=Query, mutation=Mutation)

# Create FastAPI app
app = FastAPI(
    title="LexicaForge API",
    description="Etymology and Cognate Graph API with advanced linguistic analysis",
    version="0.1.0"
)

# Add CORS middleware
allowed_origins = os.getenv("ALLOWED_ORIGINS", "http://localhost:3000,http://localhost:8000").split(",")
app.add_middleware(
    CORSMiddleware,
    allow_origins=allowed_origins,
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Add GraphQL route
graphql_app = GraphQLRouter(schema)
app.include_router(graphql_app, prefix="/graphql")

@app.get("/")
async def root():
    return {
        "message": "Welcome to LexicaForge API",
        "docs": "/docs",
        "graphql": "/graphql"
    } 